+++ Assets/Scripts/Game/Destructable.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class Destructable : MonoBehaviour, IDamagable
 {
-	[SerializeField] float health = 100;
+	[SerializeField] int health = 100;
 	[SerializeField] GameObject hitPrefab;
 	[SerializeField] GameObject destroyPrefab;
 
 	bool destroyed = false;
 
-	public void ApplyDamage(float damage)
+	public void ApplyDamage(int damage)
 	{
 		if (destroyed) return;
